Auth and save to database
Added anonymous authentication as a way of saving data without logging in.

Attributes, choices, path, currentChapter and name are now saved in the users collection and are retreived when the app is started.

Also added function to delete the user and authenticate a new one when the player starts a new story.